requirements.txt
azure-functions
pyodbc
pandas
azure-storage-blob
